requirements.txt
django-conneg>=0.6
urllib2_kerberos==0.1.6
kerberos==1.1.1
